Some additional functions for my use of plenv
Now exporting two function, `perl_clear_environment` and
`perl_libdir`. The second of these let you specify a directory in
which perl should look for modules. The default is `./local`. In
addition to that it adds `lib/` and `t/lib` to $PERL5LIB if those
directories exist. This mean I can now run the following when opening
a new terminal and have my nicely isolated enviroment set up for emacs
and other applications that I use with / for Perl development.

  $ perl_libdir
	$ emacs